Can close, without preview window
